requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.1
annotated-types==0.7.0
APScheduler==3.11.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Deprecated==1.3.1
distro==1.9.0
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
netaddr==1.3.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
wrapt==2.0.1
yarl==1.22.0
yookassa==3.8.0
